Updated gantry height

See merge request kennethsohyq/school/university/ntu-cz2006-underconstruction!72

+++ app/src/main/java/sg/edu/ntu/scse/cz2006/gymbuddies/data/CarPark.java
@@ -94,7 +94,7 @@ public class CarPark {
      */
     @NonNull
     @ColumnInfo(name = "gantry_height")
-    public int gantryHeight;
+    public double gantryHeight;
 
     /**
      * attribute of schema, relate to 'address' in actual table
